+++ libs/gltfio/src/AssetLoader.cpp
@@ -60,33 +60,6 @@ namespace gltfio {
 
 static const auto FREE_CALLBACK = [](void* mem, size_t, void*) { free(mem); };
 
-// MeshCache
-// ---------
-// If a given glTF mesh is referenced by multiple glTF nodes, then it generates a separate Filament
-// renderable for each of those nodes. All renderables generated by a given mesh share a common set
-// of VertexBuffer and IndexBuffer objects. To achieve the sharing behavior, the loader maintains a
-// small cache. The cache keys are glTF mesh definitions and the cache entries are lists of
-// primitives, where a "primitive" is a reference to a Filament VertexBuffer and IndexBuffer.
-struct Primitive {
-    VertexBuffer* vertices = nullptr;
-    IndexBuffer* indices = nullptr;
-    Aabb aabb; // object-space bounding box
-};
-using MeshCache = tsl::robin_map<const cgltf_mesh*, std::vector<Primitive>>;
-
-// MatInstanceCache
-// ----------------
-// Each glTF material definition corresponds to a single filament::MaterialInstance, which are
-// cached here in the loader. The filament::Material objects that are used to create instances are
-// cached in MaterialProvider. If a given glTF material is referenced by multiple glTF meshes, then
-// their corresponding filament primitives will share the same Filament MaterialInstance and UvMap.
-// The UvMap is a mapping from each texcoord slot in glTF to one of Filament's 2 texcoord sets.
-struct MaterialEntry {
-    MaterialInstance* instance;
-    UvMap uvmap;
-};
-using MatInstanceCache = tsl::robin_map<intptr_t, MaterialEntry>;
-
 // Sometimes a glTF bufferview includes unused data at the end (e.g. in skinning.gltf) so we need to
 // compute the correct size of the vertex buffer. Filament automatically infers the size of
 // driver-level vertex buffers from the attribute data (stride, count, offset) and clients are
@@ -171,10 +144,8 @@ struct FAssetLoader : public AssetLoader {
     MaterialProvider* mMaterials;
     Engine* mEngine;
 
-    // The loader owns a few transient mappings used only for the current asset being loaded.
+    // Transient state used only for the asset currently being loaded:
     FFilamentAsset* mResult;
-    MatInstanceCache mMatInstanceCache;
-    MeshCache mMeshCache;
     const char* mDefaultNodeName;
     bool mError = false;
     bool mDiagnosticsEnabled = false;
@@ -280,8 +251,8 @@ void FAssetLoader::createAsset(const cgltf_data* srcAsset, size_t numInstances) 
             createEntity(nodes[i], mResult->mRoot, true, nullptr);
         }
     } else {
-        // Create a separate entity hierarchy for each instance. Note that mMeshCache (vertex
-        // buffers and index buffers) and mMatInstanceCache (materials and textures) help avoid
+        // Create a separate entity hierarchy for each instance. Note that MeshCache (vertex
+        // buffers and index buffers) and MatInstanceCache (materials and textures) help avoid
         // needless duplication of resources.
         for (size_t index = 0; index < numInstances; ++index) {
             // Create a root node within each instance that is a child of the primary root.
@@ -329,9 +300,6 @@ void FAssetLoader::createAsset(const cgltf_data* srcAsset, size_t numInstances) 
         mResult->mResourceUris.push_back(pair.second);
     }
 
-    // We're done with the import, so free up transient bookkeeping resources.
-    mMatInstanceCache.clear();
-    mMeshCache.clear();
     mError = false;
 }
 
@@ -406,11 +374,11 @@ void FAssetLoader::createRenderable(const cgltf_node* node, Entity entity, const
     // If the mesh is already loaded, obtain the list of Filament VertexBuffer / IndexBuffer objects
     // that were already generated (one for each primitive), otherwise allocate a new list of
     // pointers for the primitives.
-    auto iter = mMeshCache.find(mesh);
-    if (iter == mMeshCache.end()) {
-        mMeshCache[mesh].resize(nprims);
+    auto iter = mResult->mMeshCache.find(mesh);
+    if (iter == mResult->mMeshCache.end()) {
+        mResult->mMeshCache[mesh].resize(nprims);
     }
-    Primitive* outputPrim = mMeshCache[mesh].data();
+    Primitive* outputPrim = mResult->mMeshCache[mesh].data();
     const cgltf_primitive* inputPrim = &mesh->primitives[0];
 
     Aabb aabb;
@@ -849,8 +817,8 @@ void FAssetLoader::createCamera(const cgltf_camera* camera, Entity entity) {
 MaterialInstance* FAssetLoader::createMaterialInstance(const cgltf_material* inputMat,
         UvMap* uvmap, bool vertexColor) {
     intptr_t key = ((intptr_t) inputMat) ^ (vertexColor ? 1 : 0);
-    auto iter = mMatInstanceCache.find(key);
-    if (iter != mMatInstanceCache.end()) {
+    auto iter = mResult->mMatInstanceCache.find(key);
+    if (iter != mResult->mMatInstanceCache.end()) {
         *uvmap = iter->second.uvmap;
         return iter->second.instance;
     }
@@ -1074,7 +1042,7 @@ MaterialInstance* FAssetLoader::createMaterialInstance(const cgltf_material* inp
         }
     }
 
-    mMatInstanceCache[key] = {mi, *uvmap};
+    mResult->mMatInstanceCache[key] = {mi, *uvmap};
     return mi;
 }
