+++ libs/gltfio/src/FFilamentAsset.h
@@ -28,6 +28,8 @@
 #include <filament/TransformManager.h>
 #include <filament/VertexBuffer.h>
 
+#include <gltfio/MaterialProvider.h>
+
 #include <math/mat4.h>
 
 #include <utils/Entity.h>
@@ -72,6 +74,33 @@ struct TextureSlot {
     filament::TextureSampler sampler;
     bool srgb;
 };
+
+// MeshCache
+// ---------
+// If a given glTF mesh is referenced by multiple glTF nodes, then it generates a separate Filament
+// renderable for each of those nodes. All renderables generated by a given mesh share a common set
+// of VertexBuffer and IndexBuffer objects. To achieve the sharing behavior, the loader maintains a
+// small cache. The cache keys are glTF mesh definitions and the cache entries are lists of
+// primitives, where a "primitive" is a reference to a Filament VertexBuffer and IndexBuffer.
+struct Primitive {
+    filament::VertexBuffer* vertices = nullptr;
+    filament::IndexBuffer* indices = nullptr;
+    filament::Aabb aabb; // object-space bounding box
+};
+using MeshCache = tsl::robin_map<const cgltf_mesh*, std::vector<Primitive>>;
+
+// MatInstanceCache
+// ----------------
+// Each glTF material definition corresponds to a single filament::MaterialInstance, which are
+// temporarily cached during loading. The filament::Material objects that are used to create instances are
+// cached in MaterialProvider. If a given glTF material is referenced by multiple glTF meshes, then
+// their corresponding filament primitives will share the same Filament MaterialInstance and UvMap.
+// The UvMap is a mapping from each texcoord slot in glTF to one of Filament's 2 texcoord sets.
+struct MaterialEntry {
+    filament::MaterialInstance* instance;
+    UvMap uvmap;
+};
+using MatInstanceCache = tsl::robin_map<intptr_t, MaterialEntry>;
 
 struct FFilamentAsset : public FilamentAsset {
     FFilamentAsset(filament::Engine* engine, utils::NameComponentManager* names,
@@ -218,6 +247,8 @@ struct FFilamentAsset : public FilamentAsset {
     const cgltf_data* mSourceAsset = nullptr;
     NodeMap mNodeMap; // unused for instanced assets
     std::vector<std::pair<const cgltf_primitive*, filament::VertexBuffer*> > mPrimitives;
+    MatInstanceCache mMatInstanceCache;
+    MeshCache mMeshCache;
 };
 
 FILAMENT_UPCAST(FilamentAsset)
